Extract RubyClassFile to its own file
